objects: renamed InstallationSetManager to ObjectsManager
