+++ tests/objects/fixtures/activeinactive-mode.txt
