+++ tests/objects/fixtures/single-mode.txt
