+++ tests/objects/test_manager.py
@@ -5,9 +5,8 @@ import os
 import unittest
 from unittest.mock import Mock, patch
 
-from uhu.core import manager
-from uhu.core.manager import (
-    InstallationSet, InstallationSetManager, InstallationSetMode)
+from uhu.core.objects import (
+    InstallationSet, ObjectsManager, InstallationSetMode)
 
 
 def verify_all_modes(fn):
@@ -32,7 +31,7 @@ class InstallationSetModeTestCase(unittest.TestCase):
             InstallationSetMode.from_objects([])
 
 
-class InstallationSetManagerTestCase(unittest.TestCase):
+class ObjectsManagerTestCase(unittest.TestCase):
 
     def setUp(self):
         self.options = {
@@ -43,13 +42,13 @@ class InstallationSetManagerTestCase(unittest.TestCase):
 
     @verify_all_modes
     def test_can_create_managers(self, mode):
-        manager = InstallationSetManager(mode)
+        manager = ObjectsManager(mode)
         self.assertEqual(manager.mode, mode)
         self.assertEqual(len(manager), mode.value)
 
     @verify_all_modes
     def test_installation_set_as_metadata(self, mode):
-        manager = InstallationSetManager(mode)
+        manager = ObjectsManager(mode)
         manager.create('raw', self.options)
         metadata = manager.metadata()
         self.assertEqual(len(metadata), mode.value)
@@ -58,7 +57,7 @@ class InstallationSetManagerTestCase(unittest.TestCase):
 
     @verify_all_modes
     def test_installation_set_as_template(self, mode):
-        manager = InstallationSetManager(mode)
+        manager = ObjectsManager(mode)
         manager.create('raw', self.options)
         template = manager.template()
         self.assertEqual(len(template), mode.value)
@@ -70,7 +69,7 @@ class InstallationSetManagerTestCase(unittest.TestCase):
         self.addCleanup(os.chdir, os.getcwd())
         fixtures_dir = 'fixtures'
         os.chdir(os.path.join(os.path.dirname(__file__), fixtures_dir))
-        manager = InstallationSetManager(mode)
+        manager = ObjectsManager(mode)
         self.options['filename'] = 'manager.txt'
         manager.create('raw', self.options)
         fn = '{}-mode.txt'.format(mode.name.lower())
@@ -78,28 +77,28 @@ class InstallationSetManagerTestCase(unittest.TestCase):
             self.assertEqual(str(manager), fp.read())
 
     def test_is_single_returns_True_when_single(self):
-        manager = InstallationSetManager(InstallationSetMode.Single)
+        manager = ObjectsManager(InstallationSetMode.Single)
         self.assertTrue(manager.is_single())
 
     def test_is_single_returns_False_when_not_single(self):
-        manager = InstallationSetManager(InstallationSetMode.ActiveInactive)
+        manager = ObjectsManager(InstallationSetMode.ActiveInactive)
         self.assertFalse(manager.is_single())
 
     @verify_all_modes
     def test_can_get_installation_set(self, mode):
-        manager = InstallationSetManager(mode)
+        manager = ObjectsManager(mode)
         installation_set = manager.get_installation_set(0)
         self.assertIsInstance(installation_set, InstallationSet)
 
     @verify_all_modes
     def test_get_installation_set_raises_error_if_invalid_index(self, mode):
-        manager = InstallationSetManager(mode)
+        manager = ObjectsManager(mode)
         with self.assertRaises(ValueError):
             manager.get_installation_set(100)
 
     @verify_all_modes
     def test_can_create_object(self, mode):
-        manager = InstallationSetManager(mode)
+        manager = ObjectsManager(mode)
         self.assertEqual(len(manager.all()), 0)
         index = manager.create('raw', self.options)
         self.assertEqual(len(manager.all()), mode.value * 1)
@@ -111,7 +110,7 @@ class InstallationSetManagerTestCase(unittest.TestCase):
             self.assertEqual(obj['target'], '/dev/sda')
 
     def test_can_create_object_with_different_values(self):
-        manager = InstallationSetManager(InstallationSetMode.ActiveInactive)
+        manager = ObjectsManager(InstallationSetMode.ActiveInactive)
         self.assertEqual(len(manager.all()), 0)
         self.options['target'] = ('/dev/sda', '/dev/sdb')
         index = manager.create('raw', self.options)
@@ -122,7 +121,7 @@ class InstallationSetManagerTestCase(unittest.TestCase):
 
     @verify_all_modes
     def test_can_get_object(self, mode):
-        manager = InstallationSetManager(mode)
+        manager = ObjectsManager(mode)
         index = manager.create('raw', self.options)
         for set_index, installation_set in enumerate(manager):
             obj = manager.get(index=index, installation_set=set_index)
@@ -132,7 +131,7 @@ class InstallationSetManagerTestCase(unittest.TestCase):
 
     @verify_all_modes
     def test_can_update_asymmetrical_object_option(self, mode):
-        manager = InstallationSetManager(mode)
+        manager = ObjectsManager(mode)
         index = manager.create('raw', self.options)
         for set_index, installation_set in enumerate(manager):
             obj = manager.get(index=index, installation_set=set_index)
@@ -151,7 +150,7 @@ class InstallationSetManagerTestCase(unittest.TestCase):
 
     @verify_all_modes
     def test_can_update_symmetrical_object_option(self, mode):
-        manager = InstallationSetManager(mode)
+        manager = ObjectsManager(mode)
         index = manager.create('raw', self.options)
         for set_index, installation_set in enumerate(manager):
             obj = manager.get(index=index, installation_set=set_index)
@@ -166,7 +165,7 @@ class InstallationSetManagerTestCase(unittest.TestCase):
 
     @verify_all_modes
     def test_update_asymmetrical_option_raises_without_install_set(self, mode):
-        manager = InstallationSetManager(mode)
+        manager = ObjectsManager(mode)
         index = manager.create('raw', self.options)
 
         with self.assertRaises(ValueError):
@@ -178,7 +177,7 @@ class InstallationSetManagerTestCase(unittest.TestCase):
 
     @verify_all_modes
     def test_update_symmetrical_option_raises_error_if_install_set(self, mode):
-        manager = InstallationSetManager(mode)
+        manager = ObjectsManager(mode)
         self.options['count'] = 100
         index = manager.create('raw', self.options)
         with self.assertRaises(ValueError):
@@ -190,7 +189,7 @@ class InstallationSetManagerTestCase(unittest.TestCase):
 
     @verify_all_modes
     def test_can_remove_object(self, mode):
-        manager = InstallationSetManager(mode)
+        manager = ObjectsManager(mode)
         index = manager.create('raw', self.options)
         self.assertEqual(len(manager.all()), 1 * mode.value)
         manager.remove(index)
@@ -198,7 +197,7 @@ class InstallationSetManagerTestCase(unittest.TestCase):
 
     @verify_all_modes
     def test_can_get_all_objects(self, mode):
-        manager = InstallationSetManager(mode)
+        manager = ObjectsManager(mode)
         index = manager.create('raw', self.options)
         expected = [set_.get(index) for set_ in manager]
         observed = manager.all()
