+++ tests/package/test_push.py
@@ -3,7 +3,7 @@
 
 import json
 
-from uhu.core.manager import InstallationSetMode
+from uhu.core.objects import InstallationSetMode
 from uhu.core.package import Package
 from uhu.exceptions import UploadError
 from uhu.utils import SERVER_URL_VAR
