+++ uhu/core/_base.py
@@ -1,6 +1,7 @@
 # Copyright (C) 2017 O.S. Systems Software LTDA.
 # SPDX-License-Identifier: GPL-2.0
 
+
 import hashlib
 import json
 import math
